+++ src/tree-sitter/index.ts
@@ -1,0 +1,2 @@
+export * from './ExpressionBuilder.js'
+export * from './types.js'